+++ src/mlpack/core/tree/cover_tree/single_tree_traverser.hpp
@@ -51,11 +51,46 @@ class CoverTree<DistanceType, StatisticType, MatType, RootPointPolicy>::
   size_t& NumPrunes() { return numPrunes; }
 
  private:
-  //! Reference to the rules with which the tree will be traversed.
+  size_t ScaleIndex(const int scale);
+
+  // Reference to the rules with which the tree will be traversed.
   RuleType& rule;
 
-  //! The number of nodes which have been pruned during traversal.
+  // The number of nodes which have been pruned during traversal.
   size_t numPrunes;
+
+  // As we traverse, we will hold onto these MapEntry objects to track the
+  // reference nodes that need further recursion.
+  struct CoverTreeMapEntry
+  {
+    //! The node this entry refers to.
+    CoverTree<DistanceType, StatisticType, MatType, RootPointPolicy>* node;
+    //! The score of the node.
+    double score;
+    //! The base case evaluation.
+    double baseCase;
+
+    //! Comparison operator.
+    bool operator<(const CoverTreeMapEntry& other) const
+    {
+      return (score < other.score);
+    }
+  };
+
+  // Auxiliary data structures used during search.  We keep them at the class
+  // level, because the memory allocated with them can be reused for each query
+  // point.
+
+  // The scale levels for each of the eight "hot" vectors of reference nodes.
+  arma::ivec::fixed<8> hotScaleLevels;
+  // Eight vectors of "hot" reference nodes; these will contain the next levels
+  // to be recursed into.
+  std::vector<CoverTreeMapEntry> hotScaleVectors[8];
+  // A "cold" map of reference nodes with scales that are less than the smallest
+  // scale contained in any of the hot vectors.
+  std::map<int, std::vector<CoverTreeMapEntry>, std::greater<int>> mapQueue;
+  // A vector of leaves (reference nodes with scale INT_MIN).
+  std::vector<CoverTreeMapEntry> leafVector;
 };
 
 } // namespace mlpack